requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1

python-multipart==0.0.12
requests==2.32.3

PyJWT==2.9.0

openai==1.57.4
